requirements.txt
asgiref==3.8.1
Django==5.0.7
djangorestframework==3.15.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.1
pillow==10.4.0

django-filter~=24.2
stripe~=10.5.0
celery~=5.4.0
django-celery-beat==2.6.0
